+++ gravitee-node-services/gravitee-node-services-initializer/src/main/java/io/gravitee/node/upgrader/spring/InitializerConfiguration.java
@@ -17,9 +17,7 @@ package io.gravitee.node.upgrader.spring;
 
 import io.gravitee.common.component.LifecycleComponent;
 import io.gravitee.node.api.service.InitializerService;
-import io.gravitee.node.api.service.UpgraderService;
 import io.gravitee.node.upgrader.InitializerServiceImpl;
-import io.gravitee.node.upgrader.UpgraderServiceImpl;
 import java.util.ArrayList;
 import java.util.List;
 import org.springframework.context.annotation.Bean;
@@ -30,16 +28,11 @@ import org.springframework.context.annotation.Configuration;
  * @author GraviteeSource Team
  */
 @Configuration
-public class UpgraderConfiguration {
+public class InitializerConfiguration {
 
   @Bean
   public InitializerService initializerService() {
     return new InitializerServiceImpl();
-  }
-
-  @Bean
-  public UpgraderService upgraderService() {
-    return new UpgraderServiceImpl();
   }
 
   public static List<Class<? extends LifecycleComponent>> getComponents() {
@@ -47,17 +40,8 @@ public class UpgraderConfiguration {
 
     String upgradeMode = System.getProperty("upgrade.mode");
 
-    if (upgradeMode == null || "true".equals(upgradeMode)) {
-      components.add(UpgraderService.class);
-    }
-
     if (upgradeMode == null || "false".equals(upgradeMode)) {
       components.add(InitializerService.class);
-    }
-
-    // This will avoid running Liquibase in upgrade mode
-    if ("true".equals(upgradeMode)) {
-      System.setProperty("management.jdbc.liquibase", "false");
     }
 
     return components;
